% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestIV.R
\name{ForestIV}
\alias{ForestIV}
\title{ForestIV Main Function}
\usage{
ForestIV(
  data_test,
  data_unlabel,
  control,
  method,
  iterative = TRUE,
  ntree,
  model_unbias,
  family,
  diagnostic,
  select_method
)
}
\arguments{
\item{data_test}{Testing dataframe for random forest, must have a column named "actual" that contains the ground truth, and all trees' predictions.}

\item{data_unlabel}{Unlabel dataframe for random forest, must have all trees' predictions.}

\item{control}{A character vector of control variable names. Pass an empty vector if there are no control variables}

\item{method}{"Lasso" for ForestIV method and "IIV" for EnsembleIV method.}

\item{iterative}{Whether to perform iterative IV selection or not, default to TRUE. Only relevant when method = "Lasso"}

\item{ntree}{Number of trees in the random forest.}

\item{model_unbias}{Unbiased estimation.}

\item{family}{Model specification, same as in the family parameter in glm.}

\item{diagnostic}{Whether to output diagnostic correlations for instrument validity and strength, default to TRUE.}

\item{select_method}{method of IV selection. One of "optimal" (LASSO based), "top3", and "PCA".}
}
\value{
ForestIV estimation results
}
\description{
This function implements the main ForestIV approach.
}
