% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestIV.R
\name{ForestIV}
\alias{ForestIV}
\title{ForestIV Main Function}
\usage{
ForestIV(
  data_test,
  data_unlabel,
  control,
  iterative = TRUE,
  ntree,
  model_unbias,
  diagnostic
)
}
\arguments{
\item{data_test}{Testing dataframe for random forest, must have a column named "actual" that contains the ground truth, and all trees' predictions.}

\item{data_unlabel}{Unlabel dataframe for random forest, must have all trees' predictions.}

\item{control}{A character vector of control variable names.}

\item{iterative}{Whether to perform iterative IV selection or not, default to TRUE.}

\item{ntree}{Number of trees in the random forest.}

\item{model_unbias}{A lm object of unbiased estimation.}

\item{diagnostic}{Whether to output diagnostic correlations for instrument validity and strength, default to TRUE.}
}
\value{
ForestIV estimation results
}
\description{
This function implements the main ForestIV approach.
}
